Adapted the interaction with the cache and added tests for the cache
